Moved demonstration "test" to doc-examples
